requirements.txt
jupyterlab
numpy
matplotlib-venn
pandas
tqdm
xmltodict
SPARQLWrapper
biopython >= 1.79